requirements.txt
Flask==2.0.1
flask-sqlalchemy==2.5.1
flask-login==0.6.2
flask-migrate==4.0.4
python-dotenv==1.0.0
requests==2.31.0
openai==1.3.5
python-pptx==0.6.21
Pillow==10.0.0
gunicorn==21.2.0
